Move tests into existing FolderPluginTest
